initial commit
Create a python script to (eventually) convert an ODF to a textual
representation. The goal of odf2txt is to have more options than odt2txt
and to do a better job of extracting content from other document types
(at least spreadsheets). In particular, the objective is to produce a
textual representation that forms a reasonable basis for comparing
different revisions of ODF documents (i.e. it is intended to be useful
as a textconv for git diffs).

Besides raw text output, odf2txt will partially support HTML/CSS-like
style markup (when --style is specified). The goal is NOT to reimplement
odf2xhtml, but to provide enough formatting information to identify
interesting changes, without polluting the output with uninteresting
markup that LO/OO likes to change at the drop of a hat (which is not
helpful for comparing documents).

To this end, the current state of the tool is that it implements some
simple SAX parsing of the document's styles.xml with a context parser
stack. The actual text is not parsed at present (although there is a
--raw option to dump the unfiltered and unprettified XML, including
styles.xml). Mostly, this first commit is defining the very basic
functionality (extracting of relevant data from the ODF container and a
framework for SAX parsing with a context stack).