add README
Add an overview/history document (in reStructuredText markup) providing
the first stage of some reasonable documentation for odft2xt. This is a
bit inaccurate at the moment in that it optimistically refers to a usage
document that has not yet been written, and also improved sentence
breaking (of which not even the crude 'proof of concept' implementation
is committed).