fix output misordering
Always flush text content when writing new content at the end of an
element. I originally wanted to avoid this in the thought that it might
cause sentence breaking to mis-break, but the mixture of immediate and
delayed writing was causing content to be printed after closing tags
that should have followed it.