start implementing table support
Add parsing of table tags. This works for some simple cases, but style
parsing is not yet implemented, and newline suppression of nested
paragraph tags is rather overly effective.

(Keeping simple cells free of newlines is desired, but cells containing
multiple paragraphs also end up with no newlines, which is not. The best
fix would be to break after the <td> tag when the cell contains multiple
paragraphs, indent the paragraphs, and otherwise break normally. However
this is difficult to do with a streaming parser. For now I don't tend to
have tables with large amounts of text in each cell, so it's probably
going to be left as-is for the time being. Eventually this might be
addressed by building a partial DOM of just the table cell contents.)