make inline style rules mutable
Maintain inline style rules as a list rather than joining them
immediately. Modify extract_style_rules to append to an existing list.
This will allow extract_style_rules to edit existing rules, which is
needed to extract tab stops into a single rule, without other
non-trivial modification of the code to deal with the fact that we have
an XML stream and not a DOM.